--- Hydra_Example/ISTA/dataset_test/IndicatorBox/.hydra/overrides.yaml
@@ -1,1 +1,0 @@
-[]
