--- Hydra_Example/ISTA/dataset_test/IndicatorBox/tmp_4_1649166325381.hs
@@ -1,50 +1,0 @@
-!INTERFILE  :=
-!imaging modality := PT
-name of data file := tmp_4_1649166325381.s
-originating system := userdefined
-!version of keys := STIR4.0
-!GENERAL DATA :=
-!GENERAL IMAGE DATA :=
-!type of data := PET
-imagedata byte order := LITTLEENDIAN
-!PET STUDY (General) :=
-!PET data type := Emission
-applied corrections := {None}
-!number format := float
-!number of bytes per pixel := 4
-number of dimensions := 4
-matrix axis label [4] := segment
-!matrix size [4] := 1
-matrix axis label [3] := view
-!matrix size [3] := 64
-matrix axis label [2] := axial coordinate
-!matrix size [2] := { 1}
-matrix axis label [1] := tangential coordinate
-!matrix size [1] := 192
-minimum ring difference per segment := { -1}
-maximum ring difference per segment := { 1}
-Scanner parameters:= 
-Scanner type := userdefined
-Number of rings                          := 2
-Number of detectors per ring             := 512
-Inner ring diameter (cm)                 := 102
-Average depth of interaction (cm)        := 0.7
-Distance between rings (cm)              := 1.35
-Default bin size (cm)                    := 0.3129
-View offset (degrees)                    := 0
-Maximum number of non-arc-corrected bins := 192
-Default number of arc-corrected bins     := 192
-Number of blocks per bucket in transaxial direction         := 1
-Number of blocks per bucket in axial direction              := 1
-Number of crystals per block in axial direction             := 1
-Number of crystals per block in transaxial direction        := 8
-Number of detector layers                                   := 1
-Number of crystals per singles unit in axial direction      := 1
-Number of crystals per singles unit in transaxial direction := 8
-Scanner geometry (BlocksOnCylindrical/Cylindrical/Generic)  := Cylindrical
-end scanner parameters:=
-effective central bin size (cm) := 0.317225
-number of time frames := 1
-start vertical bed position (mm) := 0
-start horizontal bed position (mm) := 0
-!END OF INTERFILE :=
