--- Hydra_Example/__init__.py
@@ -1,3 +1,0 @@
-from cfgs import *
-from coordinators import *
-from src import *
