--- Hydra_Example/cfgs/dataset/dataset_test.yaml
@@ -1,3 +1,0 @@
-name: dataset_test
-_target_: src.dataset.dataset_test.dataset_test
-
