--- Hydra_Example/cfgs/dataset/pet.yaml
@@ -1,5 +1,0 @@
-name: dataset_test
-_target_: src.dataset.dataset_test.dataset_test
-
-
-data:
