--- Hydra_Example/cfgs/dataset/pet/dataset_test/dataset_test.yaml
@@ -1,13 +1,0 @@
-Name: dataset_test
-_target_: src.dataset.dataset_test.dataset_test
-
-Dataset:
-	noise_level: 1
-	slice: 32
-Operator:
-	Angles: [1,2,3,4,5,…,359]
-Quality_Metrics:
-	psnr: true
-	ssim: true
-	roi: true
-	mse: true
