--- Hydra_Example/cfgs/functional_2/indicator_func.yaml
@@ -1,3 +1,0 @@
-name: IndicatorBox
-_target_: src.functional_2.indicator_func.indicator_func
-
