+++ Hydra_Example/cfgs/modality/.ipynb_checkpoints/ct-checkpoint.yaml
@@ -1,0 +1,21 @@
+name: "ct"
+dataset:
+    name: "test"
+acq_model:
+    LOR: 2
+    num_subsets: 1
+warm_start:
+    name: "None"
+ground_truth:
+    name: "None"
+quality_metric:
+    obj_func: True
+    mse: False
+functionals:
+    datafit:
+        KL: True
+    functional:
+        name: "NonNeg"
+algorithm:
+    name: "ISTA"
+    preconditioning: True
