+++ Hydra_Example/src/Classes/.ipynb_checkpoints/Dataset-checkpoint.py
@@ -1,0 +1,3 @@
+
+class Dataset(object):
+    def __init(self, cfg)__:
