+++ Hydra_Example/src/Classes/Dataset.py
@@ -1,0 +1,3 @@
+
+class Dataset(object):
+    def __init(self, cfg)__:
